pyproject.toml
[project]
name = "auto-apply-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "beautifulsoup4==4.14.3",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "colorama==0.4.6",
    "distro==1.9.0",
    "docx2pdf==0.1.8",
    "dotenv==0.9.9",
    "exceptiongroup==1.3.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.13.0",
    "lxml==6.0.2",
    "openai==2.30.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-docx==1.2.0",
    "python-docx-replace==0.4.4",
    "python-dotenv==1.2.1",
    "pywin32==311",
    "regex==2026.1.15",
    "requests==2.32.5",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
]
